backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-dotenv==1.0.1
python-multipart==0.0.9
pypdf==4.3.1
python-docx==1.1.2
chromadb==0.5.5
anthropic==0.34.2
openai==1.47.0
google-generativeai==0.7.2
httpx==0.27.2
pydantic==2.9.2
aiofiles==24.1.0
